// risksmart-app-main 2/packages/shared/src/forms/issues/types.ts
export type IssueTaxonomyKeys =
  | 'issues'
  | 'issuesBreachLog'
  | 'issuesConsumerDuty'
  | 'issuesCustomerTrust'
  | 'issuesGDPRBreachLog'
  | 'issuesPCIBreachLog'
  | 'issuesRiskEvents'
  | 'issuesSARLog';

export type IssueAssessmentTaxonomyKeys =
  | 'issueAssessment'
  | 'issueBreachLogAssessment'
  | 'issueConsumerDutyAssessment'
  | 'issueCustomerTrustAssessment'
  | 'issueGDPRBreachLogAssessment'
  | 'issuePCIBreachLogAssessment'
  | 'issueRiskEventAssessment'
  | 'issueSARLogAssessment';